Success!! :D
Fixed the encoding bug by switching reading and writing of streams from InputStream and OutputStream to InputStreamReader and OutputStreamWriter.
The latter use the former, but the latter are encoding-aware, whereas the former are not. Changed over TaggerTest to use them as well.

Tagger now successfully passes both the single sentence and the 200 sentence tests! ^_^ V